Add a changelog file for the last ~2 months.
Other committers: any new features, and any major bugfixes, should be added in
here, please. Try and keep it up to date. Minor changes (particularly those
not visible to the user) and so on don't generally need an entry here - this
should be a usable summary of new stuff for the non-developer.

svn path=/trunk/; revision=4386