Moderize autotools macro syntax.
We now use a 'config.h' in the autotools build for what good it does to
shorten the compiler command lines. libtool still clutters things up,
but warnings are a little more obvious now.

This involves adding a new include to all the source files; the library
is easy because everyone includes encoder_internal.h so we just put it
there; watch this if you need any of the configure stuff in another
header. The config.h include is added explicitly to each example source
file.

Also, correct reporting logic for when portaudio is not available.

svn path=/trunk/; revision=5909