Save the limit value table for the in-loop block filter in the setup header.
THIS IS AN INCOMPATIBLE BITSTREAM CHANGE.

This table, indexed by q, is used to compute the width of the triangle
waveform in the PB_INSTANCE.FiltBoundingValue table. Since it's part of
the encoder loop, we want to be able to tweak it and include it in the
bitstream.

As we understand it, the maximum useful value for an entry is 127, so we
use 7 bits to store the unsigned values. This adds another 56 bytes to
the header. Run length encoding them would help, but we don't believe
it's worth the complexity while we have no general mechanism for this
sort of data. A simple length,value pair scheme gets this below 30
bytes for the current table; if we implement this for the q matricies
we should do it here as well.

svn path=/trunk/; revision=5920