Instead of using fixed-size tables in the setup header, record a number of bits per entry before each array. This breaks byte but since the table values are often smaller than their reasonable maximums it allows saving considerable space.
THIS IS AN INCOMPATIBLE BITSTREAM CHANGE.

However, lossless transcoding is possible, as with all the recent header
changes.

For the in-loop filter limits table, 7 bits is the maximum reasonable
value so we store the entry size as a 3 bit explicit value, so the range
is 0-7 bits. (0 is an error)

For the quantizer tables, 16 bits is a general upper bound although our
current table entries are much smaller. For all of these tables we store
the entry size minus one in 4 bits, so the range is 1-16 bits.

svn path=/trunk/; revision=5955