Fix one comment (bitrate.c no longer uses a circular packet queue)
Correct a bug-that-isn't-a-bug; the bitrate manager's setpoint gravity
code relied on subtlety that just took me a while to figure out even
though I wrote it.  Given that, I could see it causing trouble in the
future if other changes broke the subtlety, so I increased the
heavy-handed logic explicitness to bulletproof the gravity code in the
cases where there's a minimum but not a maximum and vice versa.

The only change that makes a difference: the bit reservoir is set to
the bias setpoint in initialization as opposed to always starting with
a full reservoir.  This is unrelated to above, but seemed like a good
idea.


svn path=/trunk/; revision=7185