Discontinuous encoding is now working with ogg_stream_state.
Am not sure this is the best way to accomplish it, however.

I'm thinking the addition of a function for setting discontinuous may be 
better for decode because you won't know what kind of stream it is until 
you read page0, and granulepos doesn't matter for page0, so it's ok to 
switch the behavior of granulepos handling at this point.

I'll be working on this more tomarrow.


svn path=/branches/ogg2-arc/; revision=7199