More work on discontinuous stream support.  I believe I've fixed the problem from yesterday's commit, it's now honoring packet/stream abstraction (as MikeS called it).  I have not fully tested to see if it's handling CONT packets properly, but I believe it is.
There is now a ogg_stream_setmode function instead of adding a mode 
argument to ogg_stream_create.  This was done because it doesn't change 
the API for continuous codecs and because it allows the mode to be 
changed after pageno 0 has been read, which is required for decoding.

I have not added decoding support for discontinuous streams yet.


svn path=/branches/ogg2-arc/; revision=7238