Turned OGG_DISCONT into a flag, changed the conditionals to treat it as such to make it easier to add additional modes in the future, if extra modes are ever needed.
Have not yet grokked much of the decoding side of stream.c.  Added a 
conditional for discont mode on line 490, but much of the coding style 
confuses me, such as why adv is tested twice, seperatly, without 
(apparently) any possibility of modification between the two 
conditionals on lines 504 and 527.  

Need to either consult with Monty or spend more hours grokking..


svn path=/branches/ogg2-arc/; revision=7252