This is the start of a beautiful cross-platform non-linear Ogg editor.
Currently, it is capable of little more than opening a GUI and allowing 
you to add new media to the sandbox (lower left window).  When you add 
new media it detects only the codecs in the first chain, does not 
traverse the rest of the stream to find other chains, and does not 
calculate total length.  It does not yet let you drag and drop anything.

Included are icons from KDE as well as some 16x16 icons I made for 
various Xiph codecs based on Sheryl's logos.  I hope to make larger 
versions of these logos (22x22 and 32x32) for use in other sections.

It currently depends on wxPython and py-ogg2.


svn path=/trunk/; revision=7410