Wow, alot of work tonight.
Well the infobox is working now.  It will display useful information for 
files, chains, and streams and properly calculates offsets for muxed and 
chained bitstreams when calculating these totals.

I don't have the calcs for Theora or FLAC granulepos yet.  Monday night.

It's also not caching 'cut points' (page/packet boundries), but then 
again, we don't have a project workspace yet so this is not so 
important.  Now that it's processing the streams fully, not just page0, 
it's "ready" for this additional code to be added easily.

Project window is the next priority.  


svn path=/trunk/; revision=7426