Modify reverb routing in the mixdown panels.
In previous versions of Postfish, the stereo reverb output of a
channel reverb had to be explicitly mixed into an output channel using
the direct mixdown blocks or the reverb is inaudible.  You needed to
use the direct mixdown blocks to get reverb even when otherwise only
using the stereo crossplacer to produce a mixdown.

As of now, when using the Crossplacer in a mixdown panel, a channel's
reverb A output will be mixed directly into all active 'output A'
outputs, and the channel's reverb B output likewise mixed directly
into all active 'output B' outputs.  It is mixed according to the
master delay/attenuation sliders on the attenuation panel, but it does
not use the delay/attenuation settings of the crossplacer; this way
the channel can be mixed anywhere into a stereo image and the reverb
will still present the same image of space.


svn path=/trunk/; revision=8581