I'm done.  I'm through.  I've wasted the last week on GNU Autotools!
For now on I will be using SCons for all my C projects.  It's clean, 
fast, simple, efficient, and the only downside is an extra dependency.

For the love of all that is good and holy in this world, I will never 
again use GNU Autotools for any of my projects.  For group projects 
someone else will have to suffer with the chore of Autotools setup.

I will add a "configure" and "Makefile" for this when I'm closer to 
release.  Both configure and building can be done by running "scons" in 
the project's root directory.  Simple, clean, and fast.

I'll get back to real programming after I finish this chamomile tea...


svn path=/trunk/; revision=9118