Correct a board geenration bug found by Jared Lash;
the bug caused edges from the wrong region of a graph with fully partitioned sections to be considered in edge population, causing the possibility of intersections that were not recorded int he objective...

AFAIK, it affected only board 69, but was a serious logic bug
nonetheless.  Board 69 is now demonstrated solvable (it was not
previously solvable due to the bug).


svn path=/trunk/; revision=11495