* souffleur.glade
        Anyone interested on actually running Souffleur will notice the
difference. A new GUI concept was developed. Panes can be hidden and
soon, detachable.

* Souffleur.py

        Oh dear, where should I start...

        - videowidget initial size is now set according to media information
and can be correctly resized
        - Introduce new GUI concept of hiding panes
        - Add new callbacks for improving usability of the subtitle pane
                * Double-click seeks media
                * Single click edits the subtitle
        - Prepared GUI for "edit mode" and "online mode"
                * The idea behind this is that subtitles are always done in 5 stages.
First we transcript, second translate, third spotting and forth timing
and fith review. "Editing mode" will be good for everything but
reviewing. "Online mode" will make the subtitles move according to the
movie on the TreeView and on TextView, but TextView won't be sensitive
at this time.
        - Subtitle editing is done in real-time (a few bugs need fix here)
        - New framerate calculation and display in realtime
        - Added new stream information display on Project Pane

 * Subtitles.py

        - Started to remodel so we can support other formats
        - Why should we update the whole list when editing the sub text? Added
a new method for doing that nicely.

svn path=/trunk/; revision=14686