Correct a small raft of transport errors that have crept in due to insufficient testing while keeping up with constant kernel API churn. Among other things, density det for very old drives has been broken for some time, as was command set scanning.
Correct SGIO error detection such that sense keys are being properly
parsed and then the results actually returned somewhere when errors
happen.

Add a new error return code for the cdrom going missing in the middle
of an operation (404 amusingly enough; no really, it was the next in
sequence) and propogate that error upwards through the various layers;
errno is also set to ENOMEDIUM.

Allow main application read loop to bail on such a 404.

Continue working on cache detection fun



svn path=/trunk/; revision=15177