Fix a memory corruption error.
In r15977 the ref_frame_data member of oc_theora_state was changed from
a single data buffer to an array of three such buffers, but the code
which cleared the reference frame when an inter frame was received
without a previous intra frame, as from a damaged stream, was not updated,
resulting in the memset overwriting other parts of the state structures
and subsequent memory, generally resulting in a segfault.

Instead, clear the first such buffer. Patch from Timothy Terriberry.

svn path=/branches/theora-thusnelda/; revision=16099