The code in select_qi that looks for budget underflow and force-reduces the log_qtarget was predicated on "if drop frames are enabled and not infinite-buffer two pass mode".  This was primarily driven by use cases at the time and is subtly inappropriate.
Modify the predication to 'if we don't have a minimum quality level';
this allows proper behavior in the case where we have a minimum
quality level but aren't using 'soft-target'.  Otherwise, lambda could
still drop to min when we were trying to force a minimum qi, which
results in what look like 'waterfalls' as SKIP triggers on large
number of blocks but the quantizer is still very fine.


svn path=/trunk/; revision=16484