Fix an ignored fread() warning.
GCC now warns if the return code for fread() and similar functions is 
ignored, but it can be pacified just by assigning it to a variable. 
Clang isn't pursuaded by this and still warns that nothing is done
to check the result.

This combines the dummy read of the 4 byte chunk size (which we ignore)
with the subsequent read for the 'WAVE' chunk id so that the length
check covers both.

svn path=/trunk/; revision=16517