Fix a doxygen warning about duplicate \mainpage attributes.
The original \mainpage directive in theora.h was shadowed by the newer
one in codec.h, generating a warning. This was probably reliable since
the later comes first in sort order, but was still infelicitous.

I've just commuted the second to file-level documenation. I've not
figured out how to get it on the 'Legacy pre-1.0 API' module 
documenation page, which is where it should be.

svn path=/trunk/; revision=16544