Use /bin/echo explictly in autogen.sh so -n works on MacOS X.
On MacOS X, when invoked as /bin/sh, the built-in implementation of
'echo' on GNU Bash (v3.2.17 on MacOS 10.5 and v3.2.38 on MacOS 10.6)
ignores the -n option, resulting in output like

  checking for autoconf...
  -n checking for automake 1.7 or later...
  automake

instead of the expected

  checking for autoconf...
  checking for automake 1.7 or later... automake

However, the /bin/echo command works correctly. I've not checked if
this is an issue with GNU Bash 3.2 or just Apple's versions; echo -n
works as expected in Bash 4.0.35 on my Linux machine.

I've replaced some, but not all of the invocations too keep things
readable. The only important ones are those with the -n argument.

svn path=/trunk/; revision=16759