Add directory contents to ogg123 in sorted order
Patch by Ray Kohler in this thread:
http://lists.xiph.org/pipermail/vorbis-dev/2010-March/020109.html

This patch adds a configure check for scandir and alphasort, and
if present it sorts directory contents, eg. when a directory name
is given on the commandline.

Tested on Linux with these functions present, and also tested that
the old codepath still works.

svn path=/trunk/; revision=16953