Add a tiff2theora example.
This is a modified version of png2theora, using libtiff
instead of libpng to read a sequence of tiff image files
and convert them to an Ogg Theora video.

Originally added because the Sintel film project was
giving us tiff frames and waiting for conversion to
png before encoding was tedious. In fact, it nearly
took longer than creating this program.

svn path=/trunk/; revision=17844