Numerous cleanups to keep gPlanarity up to date.
A change in GDK/cairo caching behavior (since documented, but not originally part of the API) caused cached gdk_cairo surfaces to 
become invalid between expose events.  The cairo_t must be regenerated for each event.  This fixes the behavior on some machines 
where the main window pops but is blank until it is resized.

Minor include cleanups.


svn path=/trunk/; revision=17906