Don't build the specification by default. Trac #1773.
A dependency loop in the spec makefile causes Theora.pdf
to be rebuilt even when it's up to date, such as a fresh
build from a release package. This confuses some tools.

This really just a makefile bug; I would like to find a
better fix for trunk, but for now we can work around
the problem by not (re)building the spec by default.

NB: this means one needs to './configure --enable-spec'
before trying 'make dist'.

svn path=/branches/theora-1.1/; revision=17977